tests: Split off version code into dedicated module
Version refactorings will grow version code a bit. To avoid inflating
the main module even more, we split the version code off into its own
module.
